@@ -2,4 +2,4 @@ package com.desu1243.taskforge
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity()
+class MainActivity : FlutterActivity()
